% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boot_and_pretest}
\alias{boot_and_pretest}
\title{Bootstrap Pretest}
\usage{
boot_and_pretest(
  tau,
  gamma_null,
  std,
  n,
  type = "geq",
  alpha = 0.05,
  B_min = 99,
  B_max = 12799,
  beta = 0.001,
  seed = NULL
)
}
\description{
This function performs the pretest procedure described in Davidson & MacKinnon (2000) to determine the appropriate number of simulations needed to minimize loss of power.
}
\references{
Russell Davidson & James G. MacKinnon (2000), Bootstrap tests: how many bootstraps?, Econometric Reviews, 19:1, 55-68.
}
