% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boot_pretest}
\alias{boot_pretest}
\title{Bootstrap Pretest}
\usage{
boot_pretest(
  tau,
  tau_B,
  type = "geq",
  B_max = 12799,
  alpha = 0.05,
  beta = 0.001
)
}
\description{
This function performs the pretest preocedure described in Davidson & MacKinnon (2000) to determine the appropriate number of simulations needed to minize loss of power.
}
\references{
Russell Davidson & James G. MacKinnon (2000), Bootstrap tests: how many bootstraps?, Econometric Reviews, 19:1, 55-68.
}
