% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boot_pval}
\alias{boot_pval}
\title{Compute Bootstrap P-Value}
\usage{
boot_pval(tau, tau_B, type = "geq")
}
\arguments{
\item{tau}{test statistic from observed data}

\item{tau_B}{vector of bootstrap test statistics from bootstrap samples}

\item{type}{string specifying type of test. Options are: "geq" for right-tailed test, "leq" for 
left-tailed test, "two-tailed-s" for two-tailed symmetric test and "two-tailed-ns" for two-tailed non-symmetric test.}
}
\value{
Bootstrap p-value
}
\description{
This function computes the Bootstrap p-value consistent with a left-tailed test, right-tailed test, two-tailed symmetric test, and two-tailed non-symmetric test. See ch. 4 of Davidson & MacKinnon (2004) or pages 11-12 of Davidson (2013).
}
\references{
Russell Davidson & James G. MacKinnon (2004), Econometric Theory and Methods, New York, Oxford University Press.

Russell Davidson (2013), The Bootstrap in Econometrics, CEA Conference, May 2013
}
