% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_vec}
\alias{boot_vec}
\title{Generate test statistic from bootstrap samples}
\usage{
boot_vec(param_null, std, n, B)
}
\value{
vector of simulated test statistics
}
\description{
This function simulates the DGP under the null hypothesis and computes the test statistic for each sample
}
