% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bootandBtest}
\alias{bootandBtest}
\title{Bootstrap Pretest}
\usage{
bootandBtest(
  tau,
  fn,
  Bname,
  type = "geq",
  B_min = 99,
  B_max = 12799,
  alpha = 0.05,
  beta = 0.001,
  seed = NULL,
  ...
)
}
\description{
This function performs the pretest procedure described in Davidson & MacKinnon (2000) to determine the appropriate number of simulations needed to minimize loss of power.
}
\references{
Russell Davidson & James G. MacKinnon (2000), Bootstrap tests: how many bootstraps?, Econometric Reviews, 19:1, 55-68.
}
