\name{bootBtest-package}
\alias{bootBtest-package}
\alias{bootBtest}
\docType{package}
\title{
\packageTitle{bootBtest}
}
\description{
\packageDescription{bootBtest}
This package allows the user to perform the bootstrap pre-test procedure described in Davidson & MacKinnon (2000) to determine an appropriate number of bootstrap simulations, B, to use to minimize loss of power resulting from sampling.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bootBtest}
\packageIndices{bootBtest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bootBtest}

Maintainer: \packageMaintainer{bootBtest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
