% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simu_stat}
\alias{simu_stat}
\title{Test statistic for simulated sample}
\usage{
simu_stat(y_t, x_t = NULL, null_vec, type = "F-test")
}
\description{
This function estimates the model & computes test statistics for simulated samples
}
